compiler: Use LocaType for CCompositeObject
