+++ devito/types/array.py
@@ -6,14 +6,14 @@ from sympy import Expr
 
 from devito.tools import (Reconstructable, as_tuple, c_restrict_void_p,
                           dtype_to_ctype, dtypes_vector_mapper, is_integer)
-from devito.types.basic import AbstractFunction
+from devito.types.basic import AbstractFunction, LocalType
 from devito.types.utils import CtypesFactory, DimensionTuple
 
 __all__ = ['Array', 'ArrayMapped', 'ArrayObject', 'PointerArray', 'Bundle',
            'ComponentAccess', 'Bag']
 
 
-class ArrayBasic(AbstractFunction):
+class ArrayBasic(AbstractFunction, LocalType):
 
     is_ArrayBasic = True
 
@@ -58,17 +58,17 @@ class ArrayBasic(AbstractFunction):
     def is_const(self):
         return self._is_const
 
-    @property
-    def liveness(self):
-        return self._liveness
+    # @property
+    # def liveness(self):
+    #     return self._liveness
 
-    @property
-    def _mem_internal_eager(self):
-        return self._liveness == 'eager'
+    # @property
+    # def _mem_internal_eager(self):
+    #     return self._liveness == 'eager'
 
-    @property
-    def _mem_internal_lazy(self):
-        return self._liveness == 'lazy'
+    # @property
+    # def _mem_internal_lazy(self):
+    #     return self._liveness == 'lazy'
 
 
 class Array(ArrayBasic):
